app/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic

google-generativeai
sentence-transformers
faiss-cpu

pymongo
motor

numpy
pandas
scikit-learn
tqdm

pytest
black
isort
flake8
beautifulsoup4
requests